Implement check for existing file

+++ py3/BingBackground/bing_background.py
@@ -13,11 +13,15 @@ from pyvirtualdisplay import Display
 
 
 def download_link(directory, link):
-    logging.info('Downloading %s', link)
     download_path = os.path.join(directory, os.path.basename(link))
+    # Check if file already exists
+    if os.path.exists(download_path):
+        logging.info('File {} already exists, skipping.'.format(link))
+    else:
     # Download the img
-    with urlopen(link) as image, open(download_path, 'wb') as f:
-        f.write(image.read())
+        logging.info('Downloading {}'.format(link))
+        with urlopen(link) as image, open(download_path, 'wb') as f:
+            f.write(image.read())
 
 
 def change_wp(directory, link):
@@ -52,10 +56,11 @@ def main():
     display = Display(visible=0, size=(800, 600))
     display.start()
 
+    # Launch a Firefox instance
     driver = webdriver.Firefox()
     driver.get(url)
-    logging.info('Downloading page %s...' % url)
-    time.sleep(4) #hacky sleep to allow bing homepage to load so we can grab the image
+    logging.info('Downloading page {}'.format(url))
+    time.sleep(6) #hacky sleep to allow bing homepage to load so we can grab the image
 
     # Parse the bing homepage
     soup = BeautifulSoup(driver.page_source, "html.parser")
@@ -67,6 +72,7 @@ def main():
     # Find the URL elements
     link = soup.find_all('div', {'id': 'bgDiv'})[0].attrs['style']
     img_link = link[(link.find('url("')+5):link.find('");')]
+    logging.info('Found link: {}'.format(img_link))
 
     # Download and change wallpaper
     download_link(dl_dir, img_link)
